Fix Visual Studio warnings at level /W4 (#2312)
* SharePtr fix for gcc 7.2

* Fix for Poco issue #2285

* Add missing fail() to sqlite test

* Fix for testIllegalFilePath

* Update SQLiteTest.cpp

* fixexes for Visual Studio and /W4 switch

* Fixed warnings for Foundation module

* Fixed warnings for JSON module

* Fixed warnings for XML module

* Fixed warnings for NET module

* Fix unfixable warning in VarHolder.h

* Fix Typo:

* Fixes for Net

* Fixes for Crypto

* Fixes for NetSSL

* Fixes for Zip

* Fixes for PDF

* Fixes for SQLite

* Fixes for SQLite

* Fixes for SQL

* Fixes for MongoDB

* Fixes for Redis

* Fixes for CppParser

* Fixes for CppUnit

* Fixes for PageCompiler

* Fixes for PostgreSQL

* Fixes for MySQL

* Fixes for SevenZip

* Last Warnings

* Fixes for ODBC

* fix std::type_info include

+++ Foundation/include/Poco/FPEnvironment_WIN32.h
@@ -121,7 +121,7 @@ inline bool FPEnvironmentImpl::isInfiniteImpl(long double value)
 #ifdef __MINGW32__
         return isfinite(value) == 0;
 #else
-        return _finite(value) == 0;
+        return _finite(static_cast<double>(value)) == 0;
 #endif
 }
 
@@ -151,7 +151,7 @@ inline bool FPEnvironmentImpl::isNaNImpl(long double value)
 #ifdef __MINGW32__
         return isnan(value) != 0;
 #else
-        return _isnan(value) != 0;
+        return _isnan(static_cast<double>(value)) != 0;
 #endif
 }
